Deleted old profiles and added new
The initial profiles we included were experimental, based on a plan we quickly realized wouldn't work (or would be very difficult to force to work). This version is based on a new plan to create profiles solely in RDA, which output only RDA metadata. This was drafted by Theo on 2019-04-25 based on a meeting on 2019-04-24 between Theo, Crystal, Alexis, and Ben. Changes should be committed by group members to the master branch.